Use Timer 0 for delay functionality
Previously delay functionality used Timer 1 which is already used
by system clock. It is not possible to setup OCR1B and OCR1A to
different frequency, thus it was not possible to use delay with
any other value than system clock used. Therefore, use timer 0 for
delay functionality instead so same counter register is not used.